Backend refactored to logic
