+++ src/components/Board.jsx
@@ -2,11 +2,11 @@ import React, {useMemo, useRef, useState} from "react";
 import '../css/Board.css'
 import {getCharacter, isEven} from "../Util";
 import Tile from "./Tile";
-import {isInConditionToPromote} from "../backend/piecelogic/Pawn.ts";
+import {isInConditionToPromote} from "../logic/piecelogic/Pawn.ts";
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
-import {def} from "../backend/util/ToastOptions.ts";
-import EngineCzechka from "../backend/engine/EngineCzechka.ts";
+import {def} from "../logic/util/ToastOptions.ts";
+import EngineCzechka from "../logic/engine/EngineCzechka.ts";
 
 const Board = (props) => {
     const game = props.game
