+++ src/logic/piecelogic/Bishop.ts
