+++ src/logic/util/EngineUtil.ts
